requirements.txt
bs4==0.0.1
dblpy==0.3.4
mutagen==1.44.0
pbwrap==1.2.0
python-dateutil==2.8.1
discord.py==1.3.3